Add files via upload
This files contains all the code for this project.

I have played around with using creating methods and return information to the console based on certain conditions. My code can definitely be optimised now that I have written it. But this project is the second I have done in JavaScript so I am very new to this.